Moving functions to LoggedPage
